requirements.txt
google-cloud-aiplatform[adk,agent_engines]>=1.102.0
google-cloud-storage>=2.10.0
requests>=2.31.0
python-dotenv>=1.0.1
google-adk>=1.17.0
immutabledict>=4.2.1
sqlglot>=26.10.1
db-dtypes>=1.4.2
regex>=2024.11.6
tabulate>=0.9.0
google-genai>=1.21.1
absl-py>=2.2.2
pydantic>=2.11.3
pytest>=8.3.2
